refactor: replanning the app routing
