+++ apps/web/app/(core)/(bookshelf)/made4u/_components/use-referral.tsx
@@ -1,11 +1,11 @@
 import useSWRMutation from "swr/mutation";
 
-import { useBooksContext } from "@/app/(authorized)/made4u/_components/books-context";
+import { useBooksContext } from "@/app/(core)/(bookshelf)/made4u/_components/books-context";
 import {
   getReferral,
   type Options,
   type Payload,
-} from "@/app/(authorized)/made4u/made4u-actions";
+} from "@/app/(core)/(bookshelf)/made4u/made4u-actions";
 
 export function useReferral(payload: Payload, options: Options) {
   const { books } = useBooksContext();
