+++ apps/web/app/(core)/book/[bid]/_components/author-scroll-area.tsx
