+++ apps/web/app/(core)/book/[bid]/_components/description.tsx
