+++ apps/web/app/(core)/book/[bid]/_components/publish-date.tsx
