+++ apps/web/app/(core)/book/[bid]/_components/publisher.tsx
