+++ apps/web/app/(core)/book/[bid]/_components/statistics.tsx
