+++ apps/web/app/(core)/booklist/_components/data-table-row-actions.tsx
@@ -18,8 +18,8 @@ import {
   DropdownMenuSubTrigger,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { priorities } from "@/app/(authorized)/booklist/_lib/priorities";
-import type { Booklist } from "@/app/(authorized)/booklist/page";
+import { priorities } from "@/app/(core)/booklist/_lib/priorities";
+import type { Booklist } from "@/app/(core)/booklist/page";
 
 export function DataTableRowActions<TData>({ row }: { row: Row<TData> }) {
   const booklist = row.original as Booklist;
