+++ apps/web/app/(core)/booklist/_components/data-table-toolbar.tsx
@@ -5,10 +5,10 @@ import { RxCross2 } from "react-icons/rx";
 
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { DataTableFacetedFilter } from "@/app/(authorized)/booklist/_components/data-table-faceted-filter";
-import { DataTableViewOptions } from "@/app/(authorized)/booklist/_components/data-table-view-options";
-import { priorities } from "@/app/(authorized)/booklist/_lib/priorities";
-import { statuses } from "@/app/(authorized)/booklist/_lib/statues";
+import { DataTableFacetedFilter } from "@/app/(core)/booklist/_components/data-table-faceted-filter";
+import { DataTableViewOptions } from "@/app/(core)/booklist/_components/data-table-view-options";
+import { priorities } from "@/app/(core)/booklist/_lib/priorities";
+import { statuses } from "@/app/(core)/booklist/_lib/statues";
 
 export function DataTableToolbar<TData>({ table }: { table: Table<TData> }) {
   return (
